Refactor edit event (cqrs -> app service)

+++ EventMarketplace.Application/Queries/EventQueries/GetEventsQuery.cs
@@ -9,7 +9,7 @@ public sealed class GetEventsQuery : IRequest<EventListResponse>
     public DateTime? StartDate { get; set; }
     public DateTime? EndDate { get; set; }
     public string? Title { get; set; }
-    public double? StartPrice { get; set; }
-    public double? EndPrice { get; set; }
+    public decimal? StartPrice { get; set; }
+    public decimal? EndPrice { get; set; }
     public int PageNumber { get; set; }
 }
